Cast the byte count the media browser hands to humanSize
GCC with -Wconversion refuses the implicit uint64_t to double, and the CI
build is -Werror. MSVC at this project's warning level says nothing about it,
so the local build was quiet and only CI caught it.

`humanSize` takes a double and every other caller already spells the cast out
-- DeliverPanel twice, ProjectBin once. This was the one that did not.

Also makes the nesting buffer count a constant rather than a size_t multiplied
by an int literal, for the same reason: the sign-conversion warnings that
matter here are ones no compiler on this machine emits.

+++ app/MediaBrowser.cpp
@@ -96,10 +96,10 @@ Status MediaBrowser::showFolder(const std::string& path) {
         // Folders marked rather than coloured: what somebody is scanning for
         // is "can I go in there", and a symbol survives any theme.
         auto* item = new QListWidgetItem(
-            entry.isFolder
-                ? QString("▸ %1").arg(QString::fromStdString(entry.name))
-                : QString("   %1   %2")
-                      .arg(QString::fromStdString(entry.name), chrome::humanSize(entry.bytes)),
+            entry.isFolder ? QString("▸ %1").arg(QString::fromStdString(entry.name))
+                           : QString("   %1   %2")
+                                 .arg(QString::fromStdString(entry.name),
+                                      chrome::humanSize(static_cast<double>(entry.bytes))),
             list_);
         item->setData(Qt::UserRole, QString::fromStdString(entry.path));
         item->setData(Qt::UserRole + 1, entry.isFolder);
